Refactor base code
